Google Kickstart Round C 2020 Problem 2
Edit to the Alien Piano problem. Answer is still wrong but looks more correct with the minimum incorporated. The issue may lie in that the iterated random first piano key pick is not all executed. This may be due to thee code stopping somewhere in between. I guessed this as when I print out rule break the results for the second set is only 1 and 0 but if the first piano key were to start at key 4 then the number of rule breaks would be more than one